requirements.txt
Flask==2.3.2
Pillow==10.0.0
open-clip-torch==2.16.0
torch==2.0.1
scikit-learn==1.3.0
pandas==2.1.1